+++ lib/axlsx/workbook/worksheet/row.rb
@@ -163,7 +163,7 @@ module Axlsx
       types, style, formula_values, escape_formulas, offset = options.delete(:types), options.delete(:style), options.delete(:formula_values), options.delete(:escape_formulas), options.delete(:offset)
       offset.to_i.times { |index| self[index] = Cell.new(self) } if offset
       values.each_with_index do |value, index|
-        options[:style] = style.is_a?(Array) ? style[index] : style if style
+        options[:style] = (style.is_a?(Array) ? style[index] : style) || worksheet.column_info[index]&.style
         options[:type] = types.is_a?(Array) ? types[index] : types if types
         options[:escape_formulas] = escape_formulas.is_a?(Array) ? escape_formulas[index] : escape_formulas unless escape_formulas.nil?
         options[:formula_value] = formula_values[index] if formula_values.is_a?(Array)